Updated BCF validation to be more forgiving when loading BCF files.
Now exceptions are only thrown when building BCF interactively. When loading a file the errors are appended to ValidationMessages on the BCF object - meaning the consumer can at least open the file in some form even when not adhering perfectly to the schema.

Minor refactoring to allow invalid data to be be deserialiaed after an error occurs.

Simplified some Guids validation - Guids are not nullable by default